fix(vercel): emit platform data
